requirements.txt
coremltools==6.2
matplotlib==3.7.1
numpy==1.24.2
opencv_python_headless==4.7.0.72
Pillow==9.5.0
